+++ plugin/__init__.py
@@ -1,3 +1,3 @@
 """EDPPMT: Elite Dangerous PowerPlay Merit Tracker."""
 
-__version__ = "1.14.0"
+__version__ = "1.14.1"